Cities can be formed by placing city hall, also fixed #70

+++ src/main/java/com/adnre/spacegame/util/MathHelper.java
@@ -1,11 +1,11 @@
 package com.adnre.spacegame.util;
 
 public class MathHelper {
-    public static double rotX(float angle, double x, double y){
+    public static double rotX(double angle, double x, double y){
         return (x * Math.cos(angle)) - (y * Math.sin(angle));
     }
 
-    public static double rotY(float angle, double x, double y){
+    public static double rotY(double angle, double x, double y){
         return (x * Math.sin(angle)) + (y * Math.cos(angle));
     }
